Add files via upload
Code for env models uploaded. All the folders contain the env info and they are uploaded one at a time. Once all of these are uploaded, they are joined in the object 'buoys'. Then, the prevalence data is uploaded (master spreadsheet csv) that is later merged with buoys. This new 'prev' data frame has the data combined, and the variables scaled. I removed all env info prior to our collections. I can't get the models to run, either for the group or for each state. I suspect files may be too large but at this point there may be errors or tweaks I haven't realized are necessary. Thank you!